ultimas validações inseridas e mais ajustes de performance

+++ src/main/kotlin/com/felipejdias/rinhabackend2024q1/exchange/TransactionRequest.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 data class TransactionRequest(
     @get:JsonProperty("valor")
-    val amount: Long?,
+    val amount: Double?,
     @get:JsonProperty("tipo")
     val type: String?,
     @field:JsonProperty("descricao")
